Serial transport internal header moved to included directory
